TestScriptFileInfo works and Utils creates ModuleSpecification objects
